⚡ Server finalization

See merge request EMF-Informatique/tpi/futnetsinglemaster!2

+++ 01 - Application/api/src/main/java/ch/futnetsinglemaster/api/service/impl/UtilisateurServiceIMPL.java
@@ -3,6 +3,8 @@ package ch.futnetsinglemaster.api.service.impl;
 import ch.futnetsinglemaster.api.beans.LoginRequest;
 import ch.futnetsinglemaster.api.beans.ResultJSON;
 import ch.futnetsinglemaster.api.dto.PutUserDTO;
+import ch.futnetsinglemaster.api.entity.Equipe;
+import ch.futnetsinglemaster.api.entity.Role;
 import ch.futnetsinglemaster.api.entity.Utilisateur;
 import ch.futnetsinglemaster.api.dto.PostUserDTO;
 import ch.futnetsinglemaster.api.dto.UtilisateurDto;
@@ -57,7 +59,6 @@ public class UtilisateurServiceIMPL implements UtilisateurService {
     // =====================
     @Override
     public ResultJSON saveUser(PostUserDTO u) {
-        ResultJSON resultJSON = new ResultJSON();
         Utilisateur saveUser = new Utilisateur();
         saveUser.setNom(u.getNom());
         saveUser.setPrenom(u.getPrenom());
@@ -71,44 +72,48 @@ public class UtilisateurServiceIMPL implements UtilisateurService {
         String formattedDate = myDateObj.format(myFormatObj);
         String username = u.getNom() + u.getPrenom().toUpperCase().charAt(0)+formattedDate;
         saveUser.setUsername(username);
-
         saveUser.setPassword(passwordEncoder.encode(password));
         saveUser.setIdRole(roleRepo.getReferenceById(u.getRole()));
         saveUser.setIdEquipe(equipeRepo.getReferenceById(u.getNomEquipe()));
-        Utilisateur user = userRepo.save(saveUser);
-        if (user != null){
-            Map<String, Object> model = new HashMap<>();
-            model.put("nom", user.getNom());
-            model.put("prenom", user.getPrenom());
-            model.put("username", user.getUsername());
-            model.put("password", password);
-            resultJSON = mailSender.sendSimpleEmail(user.getMail(), "Inscription FutNet Single Master", model);
-        }else {
-            resultJSON.setResponseCode(400);
-            resultJSON.setResponseTitle("User creation error");
-            resultJSON.setResponseText("L'utilisateur n'as pas été crée ou une erreur c'est produite");
+        try {
+            Utilisateur user = userRepo.save(saveUser);
+            if (user != null){
+                Map<String, Object> model = new HashMap<>();
+                model.put("nom", user.getNom());
+                model.put("prenom", user.getPrenom());
+                model.put("username", user.getUsername());
+                model.put("password", password);
+                return mailSender.sendSimpleEmail(user.getMail(), "Inscription FutNet Single Master", model);
+            }else {
+                return new ResultJSON(400 ,"User creation error", "L'utilisateur n'as pas été crée ou une erreur c'est produite");
+            }
+        }catch (Exception e){
+            return new ResultJSON(400, "User error", "Les informations entrées ne sont pas valide !");
         }
-        return resultJSON;
     }
 
     // =====================
     //         PUT
     // =====================
     @Override
-    public UtilisateurDto putUser(PutUserDTO user) {
-        ResultJSON resultJSON = new ResultJSON();
+    public ResultJSON putUser(PutUserDTO user) {
         if(userRepo.existsById(user.getId_user())){
-            Utilisateur oldUser = userRepo.getReferenceById(user.getId_user());
-            oldUser.setNom(user.getNom());
-            oldUser.setPrenom(user.getPrenom());
-            oldUser.setIdRole(roleRepo.getReferenceById(user.getRole()));
-            oldUser.setIdEquipe(equipeRepo.getReferenceById(user.getNomEquipe()));
-            oldUser.setMail(user.getMail());
+            try{
+                Utilisateur oldUser = userRepo.getReferenceById(user.getId_user());
+                oldUser.setNom(user.getNom());
+                oldUser.setPrenom(user.getPrenom());
+                oldUser.setIdRole(roleRepo.getReferenceById(user.getRole()));
+                oldUser.setIdEquipe(equipeRepo.getReferenceById(user.getNomEquipe()));
+                oldUser.setMail(user.getMail());
 
-            Utilisateur newUser = userRepo.saveAndFlush(oldUser);
-            return mapToDTOList(newUser);
+                Utilisateur newUser = userRepo.saveAndFlush(oldUser);
+                return new ResultJSON(200, "", "", mapToDTOList(newUser));
+            }catch (Exception e)
+            {
+                return new ResultJSON(400, "User error", "La modification n'as pas pus être finis", null);
+            }
         }else{
-            return null;
+            return new ResultJSON(400, "User error", "Aucun utilisateur existant", null);
         }
     }
 
@@ -117,19 +122,14 @@ public class UtilisateurServiceIMPL implements UtilisateurService {
     // =====================
     @Override
     public ResultJSON deleteUser(int userId) {
-        ResultJSON resultJSON = new ResultJSON();
+        if(!userRepo.existsById(userId))
+            return new ResultJSON(400, "User error", "L'utilisateur n'existe pas!");
         userRepo.deleteById(userId);
         if(userRepo.existsById(userId)){
-
-            resultJSON.setResponseCode(400);
-            resultJSON.setResponseTitle("User Error");
-            resultJSON.setResponseText("L'utilisateur n'as pas été supprimé !");
+            return new ResultJSON(400, "User error", "L'utilisateur n'as pas été supprimé !");
         }else{
-            resultJSON.setResponseCode(200);
-            resultJSON.setResponseTitle("Success");
-            resultJSON.setResponseText("L'utilisateur as été supprimé !");
+            return new ResultJSON(200, "Success", "L'utilisateur as été supprimé !");
         }
-        return resultJSON;
     }
